Move filters params to core
