+++ lib/assets/javascripts/new-dashboard/core/filters.js
@@ -4,7 +4,7 @@ const defaultParams = {
   shared: 'no',
   locked: false,
   only_liked: false,
-  types: 'table',
+  types: 'derived',
   deepInsights: false
 };
