+++ lib/assets/javascripts/new-dashboard/router/index.js
@@ -8,7 +8,7 @@ import Home from 'new-dashboard/pages/Home';
 import Solutions from 'new-dashboard/pages/Solutions';
 import Maps from 'new-dashboard/pages/Maps';
 import Data from 'new-dashboard/pages/Data';
-import { isAllowed } from '../store/datasets/filters';
+import { isAllowed } from '../core/filters';
 
 Vue.use(Router);
