+++ lib/assets/javascripts/new-dashboard/store/datasets/index.js
@@ -1,6 +1,6 @@
 import CartoNode from 'carto-node';
 import toObject from 'new-dashboard/utils/to-object';
-import Filters, { defaultParams as filtersDefaultParams } from 'new-dashboard/store/datasets/filters';
+import Filters, { defaultParams as filtersDefaultParams } from 'new-dashboard/core/filters';
 
 const client = new CartoNode.AuthenticatedClient();
 
@@ -59,6 +59,7 @@ const datasets = {
 
       const params = {
         ...Filters[context.state.filterType],
+        types: 'table',
         page: context.state.page,
         order: context.state.order
       };
