--- lib/assets/test/spec/new-dashboard/unit/specs/fixtures/filters.js
@@ -1,69 +1,0 @@
-const defaultParams = {
-  exclude_shared: false,
-  per_page: 2,
-  shared: 'no',
-  locked: false,
-  only_liked: false,
-  types: 'table',
-  deepInsights: false
-};
-
-const mine = {
-  ...defaultParams
-};
-
-const locked = {
-  ...defaultParams,
-  'locked': true
-};
-
-const shared = {
-  ...defaultParams,
-  'shared': 'only'
-};
-
-const favorited = {
-  ...defaultParams,
-  'only_liked': true
-};
-
-const publicPrivacy = {
-  ...defaultParams,
-  'privacy': 'public'
-};
-
-const linkPrivacy = {
-  ...defaultParams,
-  'privacy': 'link'
-};
-
-const passwordPrivacy = {
-  ...defaultParams,
-  'privacy': 'password'
-};
-
-const privatePrivacy = {
-  ...defaultParams,
-  'privacy': 'private'
-};
-
-const filters = {
-  mine,
-  locked,
-  shared,
-  favorited,
-  public: publicPrivacy,
-  link: linkPrivacy,
-  password: passwordPrivacy,
-  private: privatePrivacy
-};
-
-const allowedFilters = Object.keys(filters);
-
-export default filters;
-
-export { defaultParams };
-
-export function isAllowed (filter) {
-  return allowedFilters.includes(filter);
-}
