+++ lib/assets/test/spec/new-dashboard/unit/specs/store/datasets/datasets.spec.js
@@ -1,5 +1,4 @@
 import datasets from '../../fixtures/datasets';
-import { defaultParams as filtersDefaultParams } from 'new-dashboard/store/datasets/filters';
 import DatasetsStore from 'new-dashboard/store/datasets';
 import toObject from 'new-dashboard/utils/to-object';
 import { testAction } from '../helpers';
@@ -84,7 +83,7 @@ describe('DatasetsStore', () => {
         page: 1,
         numPages: 1,
         metadata: {
-          total_entries: 11
+          total_entries: 25
         }
       };
 
@@ -94,9 +93,9 @@ describe('DatasetsStore', () => {
 
       expect(state).toEqual({
         page: 3,
-        numPages: 6,
+        numPages: 3,
         metadata: {
-          total_entries: 11
+          total_entries: 25
         }
       });
     });
